long int

+++ rebdev.pavel/P4/main.cpp
@@ -29,7 +29,7 @@ int main(int argc, char ** argv)
     return 2;
   }
 
-  int rows = 0, colums = 0;
+  long int rows = 0, colums = 0;
   try
   {
     inputFile >> rows;
@@ -39,6 +39,7 @@ int main(int argc, char ** argv)
     }
 
     inputFile >> colums;
+
     if (!inputFile)
     {
       return 2;
@@ -56,7 +57,7 @@ int main(int argc, char ** argv)
     return (((rows || colums) == 0) ? 0 : 2);
   }
 
-  if (rows > std::numeric_limits<int>::max()/colums)
+  if (rows > std::numeric_limits<long int>::max()/colums)
   {
     inputFile.close();
     //std::cerr << "The array is too large!";
